// src/types/allProducts.type.ts
export interface IMarket {
  market: string;
  division: string;
  min: number;
  max: number;
}

export interface IPriceChange {
  dir: "up" | "down";
  pct: number;
}

export interface IProduct {
  id: number;
  slug: string;
  nameBn: string;
  category: string;
  categoryNameBn: string;
  categoryIcon: string;
  unit: string;
  image: string;
  today: number;
  yesterday: number;
  lastWeek: number;
  lastMonth: number;
  change: IPriceChange;
  markets: IMarket[];
}